refactor : image/domain 폴더 밑으로 이동
