+++ src/main/java/com/kakaotech/team14backend/inner/image/usecase/CreateImage.java
@@ -2,7 +2,7 @@ package com.kakaotech.team14backend.inner.image.usecase;
 
 import com.kakaotech.team14backend.common.FileUtils;
 import com.kakaotech.team14backend.common.UploadFileDTO;
-import com.kakaotech.team14backend.inner.image.model.Image;
+import com.kakaotech.team14backend.image.domain.Image;
 import com.kakaotech.team14backend.inner.image.repository.ImageRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
