+++ src/main/java/com/kakaotech/team14backend/post/dto/CreatePostDTO.java
@@ -1,6 +1,6 @@
 package com.kakaotech.team14backend.post.dto;
 
-import com.kakaotech.team14backend.inner.image.model.Image;
+import com.kakaotech.team14backend.image.domain.Image;
 import com.kakaotech.team14backend.inner.member.model.Member;
 
 public record CreatePostDTO(Image image, UploadPostRequestDTO uploadPostRequestDTO, Member member) {
